Fix compiler warnings
--HG--
branch : rocket-tube

+++ ClothSimulation/CLagrangianSimulation.cpp
@@ -154,9 +154,9 @@ void CLagrangianSimulation::UpdateSceneObjects()
 	{
 		//if (Particle->IsShip)
 		{
-			Particle->ExhaustObject->Settings.EmitCount = (int) (3.f + 5.f / 0.3f * Particle->Thrust);
-			Particle->ExhaustObject->Settings.MinSize = (0.6f + 0.3f / 0.3f * Particle->Thrust) * 0.8f * 0.006f;
-			Particle->ExhaustObject->Settings.MaxSize = (0.6f + 0.3f / 0.3f * Particle->Thrust) * 3.2f * 0.006f;
+			Particle->ExhaustObject->Settings.EmitCount = (int) (3.f + 5.f / 0.3f * (float) Particle->Thrust);
+			Particle->ExhaustObject->Settings.MinSize = (0.6f + 0.3f / 0.3f * (float) Particle->Thrust) * 0.8f * 0.006f;
+			Particle->ExhaustObject->Settings.MaxSize = (0.6f + 0.3f / 0.3f * (float) Particle->Thrust) * 3.2f * 0.006f;
 		}
 
 		Particle->MeshObject->SetPosition(QToCartesian(Particle->Position));
